refactor(*): names components files and references
