--- src/components/Gif.module.css
@@ -1,79 +1,0 @@
-.gif-card {
-	position: relative;
-	color: inherit;
-	display: grid;
-}
-.gif-card:nth-child(5n + 1) {
-	background-color: var(--brand-color_1);
-}
-.gif-card:nth-child(5n + 2) {
-	background-color: var(--brand-color_2);
-}
-.gif-card:nth-child(5n + 3) {
-	background-color: var(--brand-color_3);
-}
-.gif-card:nth-child(5n + 4) {
-	background-color: var(--brand-color_4);
-}
-.gif-card:nth-child(5n + 5) {
-	background-color: var(--brand-color_5);
-}
-.gif-card:nth-child(11n) {
-	/* opacity: 0.1; */
-	grid-column-end: span 2;
-	grid-row-end: span 2;
-}
-
-.gif-card:nth-child(8n) {
-	/* opacity: 0.1; */
-	grid-column-end: span 2;
-	grid-row-end: span 1;
-}
-
-.gif-card:nth-child(11n + 4) {
-	grid-column: span 2;
-	grid-row: span 1;
-}
-
-@media screen and (max-width: 45rem) {
-	.gif-card:nth-child(11n) {
-		/* opacity: 0.1; */
-		grid-column-end: span 1;
-		grid-row-end: span 1;
-	}
-	.gif-card:nth-child(8n) {
-		/* opacity: 0.1; */
-		grid-column-end: span 1;
-		grid-row-end: span 1;
-	}
-	.gif-card:nth-child(11n + 4) {
-		grid-column: span 1;
-		grid-row: span 1;
-	}
-}
-
-.gif-card_figure {
-	max-width: 100%;
-	overflow: hidden;
-	margin: 0;
-	/* max-height: 30vh; */
-}
-.gif-card_image {
-	max-width: 100%;
-	width: 100%;
-	height: 100%;
-	object-fit: cover;
-}
-
-.gif-card_description {
-	position: absolute;
-	width: 100%;
-	top: auto;
-	bottom: 0;
-	padding: 0 0.2em 0.5em;
-}
-
-.gif-card_title,
-.gif-card_sub-title {
-	margin: 0;
-}
