+++ src/components/GifList/GifList.css
@@ -1,8 +1,14 @@
+:root{
+	--gridList-column: 250px;
+	--gridList-rows: 250px;
+	--gridList-gap: 6px;
+}
+
 .gif-grid {
 	display: grid;
 	grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
 	grid-auto-rows: 210px;
-	grid-gap: 20px;
+	grid-gap: var(--gridList-gap);
 	grid-auto-flow: row dense;
 	justify-content: center;
 	padding-top: 1rem;
