+++ src/components/TrendingSearches/TrendingSearches.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 
-import Category from "../Categories";
+import Category from "components/Categories/index";
 import { getTrendingsTerms } from "service/index";
 
 export default function TrendingSearches() {
